requirements.txt
acres==0.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
attrs==25.1.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
boto3==1.36.11
botocore==1.36.11
browser-use==0.1.32
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.1.0
courlan==1.3.2
cryptography==44.0.0
databases==0.9.0
dateparser==1.2.0
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
etelemetry==0.3.1
fastapi==0.115.6
fastapi-cli==0.0.7
filelock==3.17.0
filetype==1.2.0
fireworks-ai==0.15.12
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.5.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
gotrue==2.11.4
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html2text==2024.2.26
htmldate==1.9.3
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
isodate==0.6.1
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jusText==3.0.1
langchain==0.3.14
langchain-anthropic==0.3.3
langchain-aws==0.2.12
langchain-core==0.3.33
langchain-fireworks==0.2.7
langchain-google-genai==2.0.8
langchain-ollama==0.2.2
langchain-openai==0.3.1
langchain-text-splitters==0.3.5
langsmith==0.2.11
lmnr==0.4.56
looseversion==1.3.0
lxml==5.3.0
lxml_html_clean==0.4.1
MainContentExtractor==0.0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
monotonic==1.6
multidict==6.1.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.2
ollama==0.4.7
openai==1.60.2
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-langchain==0.37.1
opentelemetry-instrumentation-requests==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
opentelemetry-instrumentation-threading==0.50b0
opentelemetry-instrumentation-urllib3==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-semantic-conventions-ai==0.4.2
opentelemetry-util-http==0.50b0
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pathlib==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
playwright==1.49.1
postgrest==0.19.3
posthog==3.11.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
prov==2.0.1
psycopg2-binary==2.9.10
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydot==3.0.4
pyee==12.0.0
Pygments==2.19.1
PyMuPDF==1.25.2
pyparsing==3.2.1
pypdf==5.2.0
pypdfium2==4.30.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.1
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
realtime==2.4.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.12.0
rsa==4.9
s3transfer==0.11.2
scipy==1.15.1
selenium==4.28.1
setuptools==75.8.0
shellingham==1.5.4
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.41.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
tenacity==9.0.0
tiktoken==0.8.0
tld==0.13
tqdm==4.67.1
trafilatura==2.0.0
traits==7.0.2
trio==0.28.0
trio-websocket==0.11.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
